2nd commit

+++ index.js
@@ -1,14 +1,15 @@
-require("dotenv").config();
 const express = require("express");
-const app = express();
-const cors = require("cors");
 const mongoose = require("mongoose");
+const cors = require("cors");
+require("dotenv").config();
 const jwt = require("jsonwebtoken");
 const bcrypt = require("bcryptjs");
+const app = express();
 const PORT = process.env.PORT || 8080;
-const MONGOURL = process.env.MONGOURL;
-
+const MONGO = process.env.MONGOURL;
 app.use(express.json());
+
+console.log("MONGO URL: ", MONGO);
 
 app.use(
   cors({
@@ -16,7 +17,7 @@ app.use(
   })
 );
 
-mongoose.connect(MONGOURL, {
+mongoose.connect(MONGO, {
   useNewUrlParser: true,
   useUnifiedTopology: true,
 });
@@ -25,7 +26,6 @@ const userSchema = new mongoose.Schema({
   username: String,
   password: String,
 });
-
 const User = mongoose.model("User", userSchema);
 
 const taskSchema = new mongoose.Schema({
@@ -34,36 +34,35 @@ const taskSchema = new mongoose.Schema({
   priority: String,
   userId: mongoose.Schema.Types.ObjectId,
 });
-
 const Task = mongoose.model("Task", taskSchema);
 
 app.post("/register", async (req, res) => {
   const { username, password } = req.body;
-  const hashed = await bcrypt(password, 10);
+  const hashed = await bcrypt.hash(password, 10);
   const user = new User({ username, password: hashed });
   await user.save();
-  res.json({ message: "User has been registered" });
+  res.json({ message: "User registered" });
 });
 
 app.post("/login", async (req, res) => {
   const { username, password } = req.body;
   const user = await User.findOne({ username });
-  if (!user || (await bcrypt.compare(password, user.password))) {
-    return res.status(401).json({ message: "Inavlid Credentials" });
+  if (!user || !(await bcrypt.compare(password, user.password))) {
+    return res.status(401).json({ message: "Invalid credentials" });
   }
-  const token = jwt.sign({ userId: user_id }, "secret", { expiresIn: "1h" });
+  const token = jwt.sign({ userId: user._id }, "secret", { expiresIn: "1h" });
   res.json({ token });
 });
 
 const authMiddleware = (req, res, next) => {
   const token = req.header("Authorization")?.replace("Bearer ", "");
-  if (!token) return res.statuse(401).json({ message: "No token" });
+  if (!token) return res.status(401).json({ message: "No token" });
   try {
-    const decode = jwt.verify(token, "secret");
-    req.userId = decode.userId;
+    const decoded = jwt.verify(token, "secret");
+    req.userId = decoded.userId;
     next();
   } catch (e) {
-    res.status(401).json({ message: "Invalid Token" });
+    res.status(401).json({ message: "Invalid token" });
   }
 };
 
@@ -79,10 +78,11 @@ app.post("/tasks", authMiddleware, async (req, res) => {
 });
 
 app.delete("/tasks/:id", authMiddleware, async (req, res) => {
-  await Task.findOneAndDelete({ _id: req.params.id.userId });
-  res.json({ message: "Task Deleted" });
+  await Task.findOneAndDelete({ _id: req.params.id, userId: req.userId });
+  res.json({ message: "Task deleted" });
 });
 
+// Update task status
 app.patch("/tasks/:id/status", authMiddleware, async (req, res) => {
   const { status } = req.body;
   const task = await Task.findOneAndUpdate(
@@ -94,7 +94,8 @@ app.patch("/tasks/:id/status", authMiddleware, async (req, res) => {
   res.json(task);
 });
 
-app.patch("/tasks/id:priority", authMiddleware, async (req, res) => {
+// Update task priority
+app.patch("/tasks/:id/priority", authMiddleware, async (req, res) => {
   const { priority } = req.body;
   const task = await Task.findOneAndUpdate(
     { _id: req.params.id, userId: req.userId },
@@ -105,4 +106,4 @@ app.patch("/tasks/id:priority", authMiddleware, async (req, res) => {
   res.json(task);
 });
 
-app.listen(PORT, () => console.log("Server is running on port:8080"));
+app.listen(PORT, () => console.log(`Server running on port ${PORT}`));
